chore: add block explorer url
