+++ libs/abstract-api-provider/src/types.ts
@@ -1,0 +1,1 @@
+export type WebbProviderType = 'web3' | 'polkadot';
