+++ libs/api-provider-environment/src/transaction/utils/index.ts
@@ -1,0 +1,1 @@
+export { default as getExplorerURI } from './getExplorerURI';
